Fix: to move dummy files under sagemaker_training
